requirements.txt
numpy==1.22.3
pandas == 1.4.3
plotly == 5.5.0
scipy == 1.7.3
streamlit == 1.10.0
scikit-learn
pyyaml
gsheetsdb==0.1.13
google-cloud-storage==2.4.0